+++ ex_1_16.c
@@ -8,7 +8,7 @@ fórmula para calcular o retorno esperado de um ativo qualquer é dada por E(Ri)
 #include <stdio.h>
 
 int main() {
-    double retorno_sem_risco = 0, retorno_mercado = 0, beta = 0;
+    float retorno_sem_risco = 0, retorno_mercado = 0, beta = 0;
 
     printf("Retorno do ativo sem risco: ");
     scanf("%lf", &retorno_sem_risco);
@@ -19,7 +19,7 @@ int main() {
     printf("Coeficiente de sensibilidade do retorno\ndo ativo a variacõoes no prêmio\nde risco de mercado: ");
     scanf("%lf", &beta);
 
-    double retorno_esperado = retorno_sem_risco + beta * (retorno_mercado - retorno_sem_risco);
+    float retorno_esperado = retorno_sem_risco + beta * (retorno_mercado - retorno_sem_risco);
 
     printf("retorno esperado: %lf\n", retorno_esperado);
