+++ ex_1_9.c
@@ -6,7 +6,7 @@
 #include <stdio.h>
 
 int main() {
-    double a = 0, b = 0, c = 0;
+    float a = 0, b = 0, c = 0;
 
     printf("a: ");
     scanf("%lf", &a);
@@ -17,11 +17,11 @@ int main() {
     printf("c: ");
     scanf("%lf", &c);
 
-    double delta = sqrt( b * b - (4 * a * c));
+    float delta = sqrt( b * b - (4 * a * c));
 
     // https://pavpanchekha.com/blog/accurate-quadratic.html
-    double resultado_1 = (2 * c) / (-b + delta);
-    double resultado_2 = (2 * c) / (-b - delta);
+    float resultado_1 = (2 * c) / (-b + delta);
+    float resultado_2 = (2 * c) / (-b - delta);
 
     printf("Primeiro resultado: %lf\n", resultado_1);
     printf("Segundo resultado: %lf\n", resultado_2);
