+++ ex_3_3.c
@@ -7,7 +7,7 @@
 #define n 5
 
 int main() {
-    double numeros[n] = {0};
+    float numeros[n] = {0};
 
     int i = 0;
     for(i = 0; i < n; i++) {
